Split perflens into subfiles
